+++ server/src/game/objects/explosion.ts
@@ -34,7 +34,8 @@ export class Explosion extends ServerGameObject{
         const floor_def=Floors[floor]
         if(this.def.bullet){
             for(let i=0;i<this.def.bullet.count;i++){
-                const b=this.game.add_bullet(this.position,this.def.bullet.def,this.owner,undefined,this.source,this.layer)
+                const b=this.game.add_bullet(this.position,this.owner,undefined,this.source,this.layer)
+                b.set_definition(this.def.bullet.def)
                 b.hit_owner=true
                 b.set_direction(random.rad())
             }